requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.0.3
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.13.0
Flask-RESTful==0.3.6
flask-restful-swagger==0.20.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
nose==1.3.7
pep8==1.7.1
pluggy==0.7.1
psycopg2-binary==2.7.5
py==1.6.0
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
pytest==3.7.4
python-coveralls==2.9.1
pytz==2018.5
PyYAML==3.13
requests==2.19.1
simplejson==3.16.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
